Move ReadAllBytesTests to own file
